Added file for PyQt and for Qt
