+++ .gitignore
@@ -169,3 +169,12 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
+
+#examples
+100/
+
+.vscode/
+
+build
+
+.DS_Store
