+++ src/cpp/main.cpp
@@ -1,0 +1,10 @@
+#include <QApplication>
+#include <QPushButton>
+
+int main(int argc, char** argv){
+    QApplication app(argc, argv);
+
+    QPushButton button("hello world!");
+    button.show();
+    return app.exec();
+}
